- move rouiter's application initialization into a separate package
