+++ api/turing/database/database.go
@@ -5,9 +5,14 @@ import (
 
 	"github.com/gojek/turing/api/turing/config"
 	"github.com/golang-migrate/migrate/v4"
-	_ "github.com/golang-migrate/migrate/v4/database/postgres" // required for gomigrate
-	_ "github.com/golang-migrate/migrate/v4/source/file"
 	"github.com/jinzhu/gorm"
+
+	// required for gomigrate
+	_ "github.com/golang-migrate/migrate/v4/database/postgres"
+	_ "github.com/golang-migrate/migrate/v4/source/file"
+
+	// enable postgres dialect
+	_ "github.com/jinzhu/gorm/dialects/postgres"
 )
 
 // InitDB initialises a database connection as well as runs the migration scripts.
