+++ engines/experiment/plugin/manager/manager.go
@@ -1,9 +1,8 @@
 package manager
 
 import (
-	"github.com/gojek/turing/engines/experiment/plugin/shared"
-
 	"github.com/gojek/turing/engines/experiment/manager"
+	"github.com/gojek/turing/engines/experiment/plugin/shared"
 )
 
 // ConfigurableExperimentManager interface of an ExperimentManager, that can be configured
